Adding hook for the changeset
